pyproject.toml
[project]
name = "pma-rag"
version = "0.1.0"
description = "RAG chatbot for project management documents (PRDs, BRDs, specs, test plans)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.10.1",
    "chromadb-client>=1.5.5",
    "langchain>=1.2.13",
    "langchain-classic>=0.4.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg[binary]>=3.3.3",
    "psycopg-pool>=3.3.0",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "fastapi[standard]>=0.135.2",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.22",
    "sse-starlette>=3.3.3",
    "langsmith>=0.7.22",
    "asyncpg>=0.31.0",
    "unstructured[pdf]>=0.16.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.9.0",
]

[project.optional-dependencies]
eval = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
markers = [
    "eval: LangSmith-backed evaluation tests (require API key and uploaded datasets)",
]
